Bluetooth: Automatically select item in device selection dialog
On initialization or re-scanning, a device is automatically selected.
Devices are select according to priority (from highest to lowest):
 - Previously selected device
 - Announced name corresponding to selected vendor/model pair
 - Announced name recognized as Bluetooth enabled dive computer
 - Paired Bluetooth device
Unpaired devices are never selected.

In rare circumstances this might lead to a "fight" between user and UI
over the selected item: User starts scan and selects an item. A
higher-prioritized device is found later. Thus, this behaviour might
need some refinement.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>